+++ src/app/cart/models/cart-item.ts
@@ -1,0 +1,6 @@
+import { Product } from '../../shared/models/product';
+
+export interface CartItem {
+  product: Product;
+  count: number;
+}