+++ src/app/product/components/product/product.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 
-import { Product } from '../../models/product';
+import { Product } from '../../../shared/models/product';
 
 @Component({
   selector: 'app-product',