+++ src/app/shared/directives/highlight.directive.ts
@@ -1,0 +1,27 @@
+import { Directive, ElementRef, HostBinding, HostListener } from '@angular/core';
+
+@Directive({
+  selector: '[appHighlight]'
+})
+export class HighlightDirective {
+  @HostBinding('class')
+  attrClass = 'highlighting';
+
+  private color = 'lightgreen';
+
+  constructor(private el: ElementRef) { }
+
+  @HostListener('mouseenter')
+  onMouseEnter(): void {
+    this.highlight(this.color);
+  }
+
+  @HostListener('mouseleave')
+  leave(): void {
+    this.highlight(null);
+  }
+
+  private highlight(color: string): void {
+    this.el.nativeElement.style.backgroundColor = color;
+  }
+}